kwsys!: allow empty environment variable for WIN32
This commit changes the `env` command mode to accept empty environment
variables when CMake is running in WIN32 environment. Previously was
the variable deleted, thus giving inconsistency in the behaviour between
Window and UNIX, as UNIX allowed empty environment variables.

To delete a environment variable, the Windows user must use the
`--unset=` argument.

BREAKING CHANGE: environment variables is not deleted when assigned empty on WIN32